+++ docker-api/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 this project does NOT adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Add labels to resources that have been reserved by course sessions
+
 ## 2.18.2
 ### Fixed
 - Container name parser for test container